pyproject.toml
[tool.poetry]
name = "cratestats.io"
version = "0.1.0"
description = "Statistics on Rust Crates"
authors = ["Simon Walker <s.r.walker101@googlemail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "^0.45.0"
jinja2 = "^2.10.3"
aiofiles = "^0.4.0"
uvloop = "^0.14.0"
uvicorn = "^0.10.9"
sqlalchemy = "^1.3.12"
psycopg2-binary = "^2.8.4"
python-dotenv = "^0.10.3"
databases = "^0.2.6"
asyncpg = "^0.20.0"
starlette-prometheus = "^0.5.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pytest = "^5.3.2"
httpie = "^1.0.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
